[CHANGE] BIG

+++ models/db_brand.js
@@ -1,16 +1,16 @@
 const db_data = require('../controllers/mg_collection');
 
 exports.brand_get = async (req) => {
-    return db_data.scentre_brand.find(
-        {
-            'brand_name': req.data_type,
-            'code_number': req.data_detail
-        }.exec((err, data) => { 
-            if (!err) 
-                return data;
-            else return err; 
-        })
-    )
+    if (req.data_type == "brand_name") {
+        return db_data.scentre_brand.find(
+            {
+                'brand_name': req.data_type,
+                'code_number': req.data_detail
+            }
+        )
+    }
+
+    
 }
 
 exports.perfumer_get = async (req) => {
@@ -18,21 +18,13 @@ exports.perfumer_get = async (req) => {
         {
             'perfumer_data': req.user_id,
             'code_number': req.br_code
-        }.exec((err, data) => { 
-            if (!err) 
-                return data;
-            else return err; 
-        })
+        }
     )
 }
 
 exports.perfume_recommend = async () => {
     return db_data.scentre_brand.find(
-        {}.exec((err, data) => { 
-            if (!err) 
-                return data;
-            else return err; 
-        })
+        {}
     )
 }
 
@@ -45,9 +37,7 @@ exports.brand_post = async (req) => {
             brand_detail: req.br_detail,
             image_web: req.br_web_bg,
             image_app: req.br_app_bg,
-            perfumer_data: req.br_perfumer_list,
-            product_data: req.br_product_data,
-            payment_method: req.payment_method
+            perfumer_data: req.br_perfumer_list
         })
 
         return brand_data.save().then(
@@ -61,10 +51,26 @@ exports.brand_post = async (req) => {
 }
 
 exports.brand_delete = async (req) => {
-    return db_data.scentre_faq.deleteMany(
-        {
-            'br_code': req.br_code,
-            'perfumer_name': req.perfumer_list
+    if (type == 0) {
+        return db_data.scentre_faq.deleteMany(
+            {
+                'br_code': req.br_code
+            }, {}
+        )
+    } else {
+        let data = db_data.scentre_brand.find(
+            { 'br_code': req.br_code }
+        )
+
+        let memberData = data.perfumer_data;
+        for (let i = 0; i < memberData.length(); i++) {
+            if (memberData[i].id == req.perfumer_id) {
+                memberData.splice(i, 1);
+            }
         }
-    )
+        return db_data.scentre_brand.findByIdAndUpdate(
+            { 'br_code': req.br_code },
+            { 'perfumer_data': memberData }
+        )
+    }
 }
